+++ src/background/extension-details-view-controller.ts
